Finished

+++ Business/GrillMenuGrillingPlan.cs
@@ -4,10 +4,13 @@ using isolutions.GrillAssesment.Client.Model;
 
 namespace rebulanyum.GrillOptimizer.Business
 {
+    /// <summary>This class represents the plan generated for grilling the menu.</summary>
     public class GrillMenuGrillingPlan
     {
+        /// <summary>Rounds to be needed for grilling the whole menu on the Grill.</summary>
         public ushort Rounds { get; set; }
 
-        public double ElapsedTime { get; set; }
+        /// <summary>The total time (in seconds) passed for Grilling the whole menu.</summary>
+        public uint ElapsedTime { get; set; }
     }
 }
